chore(sourcery.yaml): update configuration for Sourcery tool
Add 'config.ini' to the ignore list to prevent Sourcery from analyzing
this configuration file. Update GitHub settings to refine the review
process by specifying different reviewers for origin and forked
repositories. Remove specific labels and add 'documentation' and
'wontfix' to the ignore list to streamline the labeling process. These
changes aim to improve the configuration management and review workflow
for the project.