docs(README.md): update changelog with release 1.3.3 details
Update the changelog in the README to include the new point release
1.3.3, which adds a debug log file "bm_logging.py" for capturing all
Brandmeister messages. This provides users with more detailed logging
capabilities for debugging purposes.

chore(bm_monitor.py): bump version to 1.3.3

Update the version number in the bm_monitor.py script to reflect the
new point release 1.3.3. This ensures consistency between the script
and the documentation, indicating the latest changes and improvements.